+++ test/case-join.sh
@@ -5,3 +5,5 @@
 ./xsvutils ./test/data/sample-join-1.tsv join  --right ./test/data/sample-join-2.tsv --right-outer > ./var/test/case-join-right-outer.tsv
 ./xsvutils ./test/data/sample-join-1.tsv join  --right ./test/data/sample-join-2.tsv --full-outer > ./var/test/case-join-full-outer.tsv
 
+./xsvutils ./test/data/zero.tsv join  --right ./test/data/sample-join-2.tsv --right-outer > ./var/test/case-join-zero.tsv
+